Kaal Chakra/backend/v1/requirements.txt
flask
flask-cors
pure-python-adb
pillow
